--- src/types/location-type.ts
@@ -1,7 +1,0 @@
-type LocationType = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-};
-
-export type { LocationType };
