--- src/types/offer-type.ts
@@ -1,32 +1,0 @@
-export type OfferType = {
-  id: string;
-  title: string;
-  description: string;
-  type: string;
-  price: number;
-  images: string[];
-  city: {
-    name: string;
-    location: {
-      latitude: number;
-      longitude: number;
-      zoom: number;
-    };
-  };
-  location: {
-    latitude: number;
-    longitude: number;
-    zoom: number;
-  };
-  goods: string[];
-  host: {
-    isPro: boolean;
-    name: string;
-    avatarUrl: string;
-  };
-  isPremium: boolean;
-  isFavorite: boolean;
-  rating: number;
-  bedrooms: number;
-  maxAdults: number;
-};
