--- src/types/review-type.ts
@@ -1,11 +1,0 @@
-export type ReviewType = {
-  id: string;
-  comment: string;
-  date: string;
-  rating: number;
-  user: {
-    name: string;
-    avatarUrl: string;
-    isPro: boolean;
-  };
-};
